+++ tests/model_setup/test__models__.py
@@ -15,7 +15,7 @@ class TestGenerateSchema(test.SimpleTestCase):
     async def asyncSetUp(self):
         await super().asyncSetUp()
         try:
-            Tortoise.apps = {}
+            Tortoise.apps = None
             Tortoise._inited = False
         except ConfigurationError:
             pass
