+++ tests/model_setup/test_bad_relation_reference.py
@@ -7,7 +7,7 @@ class TestBadRelationReferenceErrors(test.SimpleTestCase):
     async def asyncSetUp(self):
         await super().asyncSetUp()
         try:
-            Tortoise.apps = {}
+            Tortoise.apps = None
             Tortoise._inited = False
         except ConfigurationError:
             pass
