+++ tortoise/__init__.py
@@ -6,17 +6,14 @@ import logging
 import os
 import warnings
 from collections.abc import Callable, Coroutine, Iterable
-from copy import deepcopy
-from inspect import isclass
 from types import ModuleType
-from typing import Any, cast
+from typing import Any
 
 from anyio import from_thread
-from pypika_tortoise import Query, Table
-
 from tortoise.backends.base.client import BaseDBAsyncClient
 from tortoise.backends.base.config_generator import expand_db_url, generate_config
 from tortoise.connection import connections
+from tortoise.apps import Apps
 from tortoise.exceptions import ConfigurationError
 from tortoise.fields.relational import (
     BackwardFKRelation,
@@ -25,7 +22,6 @@ from tortoise.fields.relational import (
     ManyToManyFieldInstance,
     OneToOneFieldInstance,
 )
-from tortoise.filters import get_m2m_filters
 from tortoise.log import logger
 from tortoise.models import Model, ModelMeta
 from tortoise.timezone import _reset_timezone_cache
@@ -33,7 +29,7 @@ from tortoise.utils import generate_schema_for_client
 
 
 class Tortoise:
-    apps: dict[str, dict[str, type[Model]]] = {}
+    apps: Apps | None = None
     table_name_generator: Callable[[type[Model]], str] | None = None
     _inited: bool = False
 
@@ -104,9 +100,8 @@ class Tortoise:
 
         if not models:
             models = []
-            for app in cls.apps.values():
-                for model in app.values():
-                    models.append(model)
+            if cls.apps:
+                models = list(cls.apps.get_models_iterable())
 
         return {
             f"{model._meta.app}.{model.__name__}": model.describe(serializable) for model in models
@@ -114,208 +109,11 @@ class Tortoise:
 
     @classmethod
     def _init_relations(cls) -> None:
-        def get_related_model(related_app_name: str, related_model_name: str) -> type[Model]:
-            """
-            Test, if app and model really exist. Throws a ConfigurationError with a hopefully
-            helpful message. If successful, returns the requested model.
-
-            :raises ConfigurationError: If no such app exists.
-            """
-            try:
-                return cls.apps[related_app_name][related_model_name]
-            except KeyError:
-                if related_app_name not in cls.apps:
-                    raise ConfigurationError(
-                        f"No app with name '{related_app_name}' registered."
-                        f" Please check your model names in ForeignKeyFields"
-                        f" and configurations."
-                    )
-                raise ConfigurationError(
-                    f"No model with name '{related_model_name}' registered in"
-                    f" app '{related_app_name}'."
-                )
-
-        def split_reference(reference: str) -> tuple[str, str]:
-            """
-            Validate, if reference follow the official naming conventions. Throws a
-            ConfigurationError with a hopefully helpful message. If successful,
-            returns the app and the model name.
-
-            :raises ConfigurationError: If reference is invalid.
-            """
-            if len(items := reference.split(".")) != 2:  # pragma: nocoverage
-                raise ConfigurationError(
-                    f"'{reference}' is not a valid model reference Bad Reference."
-                    " Should be something like '<appname>.<modelname>'."
-                )
-            return items[0], items[1]
-
-        def init_fk_o2o_field(model: type[Model], field: str, is_o2o=False) -> None:
-            fk_object = cast(
-                "OneToOneFieldInstance | ForeignKeyFieldInstance", model._meta.fields_map[field]
-            )
-            reference = fk_object.model_name
-            if not isinstance(reference, str):
-                related_model: type[Model] = reference
-            else:
-                related_app_name, related_model_name = split_reference(reference)
-                related_model = get_related_model(related_app_name, related_model_name)
-
-            if to_field := fk_object.to_field:
-                related_field = related_model._meta.fields_map.get(to_field)
-                if not related_field:
-                    raise ConfigurationError(
-                        f'there is no field named "{to_field}" in model "{related_model_name}"'
-                    )
-                if not related_field.unique:
-                    raise ConfigurationError(
-                        f'field "{to_field}" in model "{related_model_name}" is not unique'
-                    )
-            else:
-                fk_object.to_field = related_model._meta.pk_attr
-                related_field = related_model._meta.pk
-            key_fk_object = deepcopy(related_field)
-            fk_object.to_field_instance = related_field
-            fk_object.field_type = fk_object.to_field_instance.field_type
-
-            key_field = f"{field}_id"
-            key_fk_object.reference = fk_object
-            key_fk_object.source_field = fk_object.source_field or key_field
-            for attr in ("index", "default", "null", "generated", "description"):
-                setattr(key_fk_object, attr, getattr(fk_object, attr))
-            if is_o2o:
-                key_fk_object.pk = fk_object.pk
-                key_fk_object.unique = fk_object.unique
-            else:
-                key_fk_object.pk = False
-                key_fk_object.unique = False
-            model._meta.add_field(key_field, key_fk_object)
-            fk_object.related_model = related_model
-            fk_object.source_field = key_field
-            if (backward_relation_name := fk_object.related_name) is not False:
-                if not backward_relation_name:
-                    backward_relation_name = f"{model._meta.db_table}s"
-                if backward_relation_name in related_model._meta.fields:
-                    raise ConfigurationError(
-                        f'backward relation "{backward_relation_name}" duplicates in'
-                        f" model {related_model_name}"
-                    )
-
-                fk_relation = (
-                    BackwardOneToOneRelation(
-                        model,
-                        key_field,
-                        key_fk_object.source_field,
-                        null=True,
-                        description=fk_object.description,
-                    )
-                    if is_o2o
-                    else BackwardFKRelation(
-                        model,
-                        key_field,
-                        key_fk_object.source_field,
-                        null=fk_object.null,
-                        description=fk_object.description,
-                    )
-                )
-                fk_relation.to_field_instance = fk_object.to_field_instance
-                related_model._meta.add_field(backward_relation_name, fk_relation)
-            if is_o2o and fk_object.pk:
-                model._meta.pk_attr = key_field
-
-        for app_name, app in cls.apps.items():
-            for model_name, model in app.items():
-                if model._meta._inited:
-                    continue
-                model._meta._inited = True
-                if not model._meta.db_table:
-                    model._meta.db_table = (
-                        cls.table_name_generator(model)
-                        if cls.table_name_generator
-                        else (model.__name__.lower())
-                    )
-
-                for field in sorted(model._meta.fk_fields):
-                    init_fk_o2o_field(model, field)
-
-                for field in model._meta.o2o_fields:
-                    init_fk_o2o_field(model, field, is_o2o=True)
-
-                for field in list(model._meta.m2m_fields):
-                    m2m_object = cast(ManyToManyFieldInstance, model._meta.fields_map[field])
-                    if m2m_object._generated:
-                        continue
-                    if not (backward_key := m2m_object.backward_key):
-                        backward_key = f"{model._meta.db_table}_id"
-                        if backward_key == m2m_object.forward_key:
-                            backward_key = f"{model._meta.db_table}_rel_id"
-                        m2m_object.backward_key = backward_key
-
-                    reference = m2m_object.model_name
-                    if not isinstance(reference, str):
-                        related_model: type[Model] = reference
-                    else:
-                        related_app_name, related_model_name = split_reference(reference)
-                        related_model = get_related_model(related_app_name, related_model_name)
-
-                    m2m_object.related_model = related_model
-
-                    if not (backward_relation_name := m2m_object.related_name):
-                        backward_relation_name = m2m_object.related_name = (
-                            f"{model._meta.db_table}s"
-                        )
-                    if backward_relation_name in related_model._meta.fields:
-                        raise ConfigurationError(
-                            f'backward relation "{backward_relation_name}" duplicates in'
-                            f" model {related_model_name}"
-                        )
-
-                    if not m2m_object.through:
-                        related_model_table_name = (
-                            related_model._meta.db_table or related_model.__name__.lower()
-                        )
-                        m2m_object.through = f"{model._meta.db_table}_{related_model_table_name}"
-
-                    m2m_relation = ManyToManyFieldInstance(
-                        f"{app_name}.{model_name}",
-                        m2m_object.through,
-                        forward_key=m2m_object.backward_key,
-                        backward_key=m2m_object.forward_key,
-                        related_name=field,
-                        field_type=model,
-                        description=m2m_object.description,
-                    )
-                    m2m_relation._generated = True
-                    model._meta.filters.update(get_m2m_filters(field, m2m_object))
-                    related_model._meta.add_field(backward_relation_name, m2m_relation)
+        if not cls.apps:
+            return
+        cls.apps._init_relations()
 
     @classmethod
-    def _discover_models(cls, models_path: ModuleType | str, app_label: str) -> list[type[Model]]:
-        if isinstance(models_path, ModuleType):
-            module = models_path
-        else:
-            try:
-                module = importlib.import_module(models_path)
-            except ImportError:
-                raise ConfigurationError(f'Module "{models_path}" not found')
-        discovered_models = []
-        if possible_models := getattr(module, "__models__", None):
-            try:
-                possible_models = [*possible_models]
-            except TypeError:
-                possible_models = None
-        if not possible_models:
-            possible_models = [getattr(module, attr_name) for attr_name in dir(module)]
-        for attr in possible_models:
-            if isclass(attr) and issubclass(attr, Model) and not attr._meta.abstract:
-                if attr._meta.app and attr._meta.app != app_label:
-                    continue
-                attr._meta.app = app_label
-                discovered_models.append(attr)
-        if not discovered_models:
-            warnings.warn(f'Module "{models_path}" has no models', RuntimeWarning, stacklevel=4)
-        return discovered_models
-
     @classmethod
     def init_models(
         cls,
@@ -335,35 +133,30 @@ class Tortoise:
 
         :raises ConfigurationError: If models are invalid.
         """
-        app_models: list[type[Model]] = []
-        for models_path in models_paths:
-            app_models += cls._discover_models(models_path, app_label)
+        cls.init_app(app_label, models_paths, _init_relations=_init_relations)
 
-        cls.apps[app_label] = {model.__name__: model for model in app_models}
+    @classmethod
+    def init_app(
+        cls,
+        label: str,
+        model_paths: Iterable[ModuleType | str],
+        _init_relations: bool = True,
+    ) -> dict[str, type[Model]]:
+        """
+        Early initialization of Tortoise ORM Models for a single app.
 
-        if _init_relations:
-            cls._init_relations()
+        :param label: The app label, e.g. 'models'
+        :param model_paths: Models paths to initialize
+        :param _init_relations: Whether to init relations or not
+        """
+        if not cls.apps:
+            cls.apps = Apps({}, connections, cls.table_name_generator)
+        cls.apps._table_name_generator = cls.table_name_generator
+        return cls.apps.init_app(label, model_paths, _init_relations=_init_relations)
 
     @classmethod
     def _init_apps(cls, apps_config: dict[str, dict[str, Any]]) -> None:
-        for name, info in apps_config.items():
-            try:
-                connections.get(info.get("default_connection", "default"))
-            except KeyError:
-                raise ConfigurationError(
-                    'Unknown connection "{}" for app "{}"'.format(
-                        info.get("default_connection", "default"), name
-                    )
-                )
-
-            cls.init_models(info["models"], name, _init_relations=False)
-
-            for model in cls.apps[name].values():
-                model._meta.default_connection = info.get("default_connection", "default")
-
-        cls._init_relations()
-
-        cls._build_initial_querysets()
+        cls.apps = Apps(apps_config, connections, cls.table_name_generator)
 
     @classmethod
     def _get_config_from_config_file(cls, config_file: str) -> dict:
@@ -384,15 +177,8 @@ class Tortoise:
 
     @classmethod
     def _build_initial_querysets(cls) -> None:
-        for app in cls.apps.values():
-            for model in app.values():
-                model._meta.finalise_model()
-                model._meta.basetable = Table(name=model._meta.db_table, schema=model._meta.schema)
-                basequery = model._meta.db.query_class.from_(model._meta.basetable)
-                model._meta.basequery = cast(Query, basequery)
-                model._meta.basequery_all_fields = cast(
-                    Query, basequery.select(*model._meta.db_fields)
-                )
+        if cls.apps:
+            cls.apps._build_initial_querysets()
 
     @classmethod
     async def init(
@@ -577,11 +363,13 @@ class Tortoise:
 
     @classmethod
     async def _reset_apps(cls) -> None:
-        for app in cls.apps.values():
-            for model in app.values():
-                if isinstance(model, ModelMeta):
-                    model._meta.default_connection = None
+        if not cls.apps:
+            return
+        for model in cls.apps.get_models_iterable():
+            if isinstance(model, ModelMeta):
+                model._meta.default_connection = None
         cls.apps.clear()
+        cls.apps = None
 
     @classmethod
     async def generate_schemas(cls, safe: bool = True) -> None:
